Java: API bugfix/change: costs should have been doubles not ints.
This is for relex bug #169 at 
https://github.com/opencog/relex/issues/169


git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34800 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ bindings/java/org/linkgrammar/LGConfig.java
@@ -21,7 +21,7 @@ public class LGConfig
 {
 	private int maxLinkages = 25;
 	private int maxParseSeconds = 60;
-	private int maxCost = -1;
+	private double maxCost = -1.0;
 	private boolean allowSkippedWords = true;
 	private boolean storeConstituentString = true;
 	private boolean storeDiagramString = false;
@@ -43,11 +43,11 @@ public class LGConfig
 	{
 		maxParseSeconds = m;
 	}
-	public int getMaxCost()
+	public double getMaxCost()
 	{
 		return maxCost;
 	}
-	public void setMaxCost(int m)
+	public void setMaxCost(double m)
 	{
 		maxCost = m;
 	}
